Search & Display feature - Backbone structure
This is the backbone of the search and display feature. It's using Typescript for type safety and MUI for UI styling. The properties list are curently in a Json file. Next step is to get all these info from AOS on Google Maps. For info will be added to Jira.